Communicator_conf united into one. Beware SSH now has own config structure. From now on com_conf.ssh.uid insted of com_conf.uid.
